dateparser/requirements.txt
configparser==5.0.2
connect==0.2
greenlet==1.0.0
importlib-metadata==3.7.3
mysql-connector==2.2.9
numpy==1.16.6
opencv-python==4.2.0.32
Pillow==6.2.2
PyMySQL==1.0.2
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2021.1
ratelimiter==1.2.0.post0
six==1.15.0
SQLAlchemy==1.4.0
typing-extensions==3.7.4.3
zipp==3.4.1
